Rename files hierarchical_model_static.* to static_hierarchical_linear_model.*.
