+++ src/HealthGPS/CMakeLists.txt
@@ -43,8 +43,8 @@ target_sources(
             "healthgps.h"
             "static_linear_model.cpp"
             "static_linear_model.h"
-            "hierarchical_model_static.cpp"
-            "hierarchical_model_static.h"
+            "static_hierarchical_linear_model.cpp"
+            "static_hierarchical_linear_model.h"
             "info_message.cpp"
             "info_message.h"
             "life_table.cpp"
