+++ src/HealthGPS/hierarchical_model.h
@@ -2,4 +2,4 @@
 
 #include "energy_balance_hierarchical_model.h"
 #include "energy_balance_model.h"
-#include "hierarchical_model_static.h"
+#include "static_hierarchical_linear_model.h"
